make mic audible

+++ noita-proxy/src/net/messages.rs
@@ -34,7 +34,8 @@ pub(crate) enum NetMsg {
     RespondFlagSlow(usize, bool),
     RespondFlagMoon(i32, i32, bool),
     RespondFlagStevari(i32, i32, OmniPeerId),
-    AudioData(Vec<i16>),
+    //AudioData(Vec<Vec<u8>>),
+    AudioData(Vec<f32>),
 }
 
 impl From<MessageRequest<WorldNetMessage>> for MessageRequest<NetMsg> {
